Remove a dependência `magic` do arquivo `file_utils`, move-a para `file_utils_mimetype`. Atualiza `sps_maker` para usar `file_utils_mimetype`.
