+++ packtools/file_utils.py
@@ -1,6 +1,5 @@
 import os
 import logging
-import magic
 import re
 import shutil
 import tempfile
@@ -194,22 +193,6 @@ def extract_issn_from_zip_uri(zip_uri):
     match = re.search(r'.*/ingress/packages/(\d{4}-\d{4})/.*.zip', zip_uri)
     if match:
         return match.group(1)
-
-
-def is_valid_file(file_path, check_mimetype=False):
-    if os.path.exists(file_path):
-        if check_mimetype:
-            if get_mimetype(file_path) not in ('text/plain', 'application/xml'):
-                return False
-
-        return True
-
-    return False
-
-
-def get_mimetype(file_path):
-    with open(file_path, 'rb') as fin:
-        return magic.from_buffer(fin.read(2048), mime=True)
 
 
 def get_filename(file_path):
